Related work items: #2620

+++ Cloudents.Admin2/Models/PaymentRequest.cs
@@ -13,7 +13,7 @@ namespace Cloudents.Admin2.Models
         public long UserId { get; set; }
         public long TutorId { get; set; }
 
-        public long AdminDuration { get; set; }
+        public int AdminDuration { get; set; }
 
         [Required]
         public Guid StudyRoomSessionId { get; set; }
